[refactor] #10 시설물 서비스, 컨트롤러 리팩토링
- 시설물 서비스에서 ReportStatus를 사용하도록 리팩토링함
- ReportStatus에 상태들을 리스트로 가져오는 메서드를 추가함
- FacilityResDto의 위치를 facility 디렉토리 안으로 변경함
- FacilityController에서 바뀐 Facility 객체를 사용하도록 수정함
